feat WritingView 수정 및 WritingStore 추가
