+++ Targets/Plda/Sources/Shared/Managers/FontManager.swift
@@ -11,17 +11,17 @@ import SwiftUI
 
 extension Font {
     // ExtraBold
-    static let extraBold: Font = .custom("Pretendard-ExtraBold", size: 36)
+    static let extraBold: Font = PldaFontFamily.Pretendard.extraBold.swiftUIFont(size: 36)
 
     // Bold
-    static let bold24: Font = .custom("Pretendard-Bold", size: 24)
-    static let bold16: Font = .custom("Pretendard-Bold", size: 16)
+    static let bold24: Font = PldaFontFamily.Pretendard.bold.swiftUIFont(size: 24)
+    static let bold16: Font = PldaFontFamily.Pretendard.bold.swiftUIFont(size: 16)
     
     // SemiBold
-    static let semiBold: Font = .custom("Pretendard-SemiBold", size: 20)
+    static let semiBold: Font = PldaFontFamily.Pretendard.semiBold.swiftUIFont(size: 20)
     
     // Medium
-    static let medium16: Font = .custom("Pretendard-Medium", size: 16)
-    static let medium12: Font = .custom("Pretendard-Medium", size: 12)
+    static let medium16: Font = PldaFontFamily.Pretendard.medium.swiftUIFont(size: 16)
+    static let medium12: Font = PldaFontFamily.Pretendard.medium.swiftUIFont(size: 12)
     
 }
