--- Targets/Plda/Sources/Views/TestUI.swift
@@ -1,21 +1,0 @@
-//
-//  TestUI.swift
-//  Plda
-//
-//  Created by 김채빈 on 2023/09/04.
-//  Copyright © 2023 alom.com. All rights reserved.
-//
-
-import SwiftUI
-
-struct TestUI: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct TestUI_Previews: PreviewProvider {
-    static var previews: some View {
-        TestUI()
-    }
-}
