--- Targets/Plda/Sources/Views/WritingStore.swift
@@ -1,9 +1,0 @@
-//
-//  WritingStore.swift
-//  Plda
-//
-//  Created by 김채빈 on 2023/09/04.
-//  Copyright © 2023 alom.com. All rights reserved.
-//
-
-import Foundation
